refactor: move item icon components
