+++ app/src/components/ItemLink.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import { useStore } from '../data/store'
 import { ItemId, ItemType } from '../data/types'
 import { getSerialFromId, renderDate } from '../data/util'
-import { ItemIcon } from './ItemIcon'
+import { ItemIcon } from '../features/items'
 import { isActive, ItemStatus } from './Status'
 
 export const PATHS: { [type in ItemType]: string } = {
