+++ app/src/features/items/editor/IssueEditor.tsx
@@ -1,5 +1,5 @@
 import { Form, Input, message, Modal, Select, Tag } from 'antd'
-import { IssueIcon } from '../../../components/ItemIcon'
+import { IssueIcon } from '../ItemIcon'
 import { Controller, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { IssueId, IssueStatus, NewIssue } from '../../../data/types'
