+++ app/src/features/items/editor/RiskEditor.tsx
@@ -1,5 +1,5 @@
 import { Form, Input, message, Modal, Select, Tag } from 'antd'
-import { RiskIcon } from '../../../components/ItemIcon'
+import { RiskIcon } from '../ItemIcon'
 import { Controller, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { NewRisk, RiskId, RiskStatus } from '../../../data/types'
