+++ app/src/features/items/index.ts
@@ -3,3 +3,4 @@ export * from './details'
 export * from './table'
 export * from './useNavigateToItem'
 export * from './editor'
+export * from './ItemIcon'
