+++ app/src/features/layout/AddItemButton.tsx
@@ -2,12 +2,7 @@ import { PlusOutlined } from '@ant-design/icons'
 import { Dropdown, Menu } from 'antd'
 import Button from 'antd-button-color'
 import { useEditorStore } from '../items'
-import {
-  DecisionIcon,
-  ImprovementIcon,
-  IssueIcon,
-  RiskIcon
-} from '../../components/ItemIcon'
+import { DecisionIcon, ImprovementIcon, IssueIcon, RiskIcon } from '../items'
 
 export const AddItemButton: React.FC = () => {
   const editor = useEditorStore(
