+++ app/src/features/search/HeaderSearch.tsx
@@ -3,7 +3,7 @@ import { AutoComplete, Button, Divider, Modal, Typography } from 'antd'
 import { useState } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
 import { useNavigate } from 'react-router-dom'
-import { ItemIcon } from '../../components/ItemIcon'
+import { ItemIcon } from '../items'
 import { useStore } from '../../data/store'
 import { ItemId } from '../../data/types'
 import { useNavigateToItem } from '../items'
