+++ app/src/pages/ImprovementPage.tsx
@@ -3,7 +3,7 @@ import { useStore } from '../data/store'
 import { EditOutlined } from '@ant-design/icons'
 import { IMPROVEMENT_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
-import { ImprovementIcon } from '../components/ItemIcon'
+import { ImprovementIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
 import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
