+++ app/src/pages/IssuePage.tsx
@@ -3,7 +3,7 @@ import { useStore } from '../data/store'
 import { EditOutlined, StopOutlined } from '@ant-design/icons'
 import { ISSUE_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
-import { IssueIcon } from '../components/ItemIcon'
+import { IssueIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
 import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
