+++ app/src/pages/RiskPage.tsx
@@ -3,7 +3,7 @@ import { useStore } from '../data/store'
 import { EditOutlined } from '@ant-design/icons'
 import { RISK_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
-import { RiskIcon } from '../components/ItemIcon'
+import { RiskIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
 import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
