+++ app/src/pages/decision-details.tsx
@@ -3,7 +3,7 @@ import { useStore } from '../data/store'
 import { EditOutlined } from '@ant-design/icons'
 import { DECISION_STATUS_UI } from '../components/Status'
 import { renderDate } from '../data/util'
-import { DecisionIcon } from '../components/ItemIcon'
+import { DecisionIcon } from '../features/items'
 import { PageHeader } from '../features/layout'
 import { Error404 } from '../features/ui'
 import { useParams } from 'react-router-dom'
